+++ main/java/AllampapirStrategia/KoltsegStrategia_Bank.java
@@ -3,15 +3,16 @@ package AllampapirStrategia;
 import KoltsegDecorator.IKoltseg;
 import KoltsegDecorator.Koltseg_Alap;
 import KoltsegDecorator.Koltseg_BankSzamla;
+import Portfolio.Egyenleg;
 
 public class KoltsegStrategia_Bank extends KoltsegStrategia
 {
-    private int nevErtek;
+    private long nevErtek;
     private int futamIdo;
 
     Egyenleg myEgyenleg = Egyenleg.getInstance();
 
-    public KoltsegStrategia_Bank(int nevErtek, int futamIdo)
+    public KoltsegStrategia_Bank(long nevErtek, int futamIdo)
     {
         this.nevErtek = nevErtek;
         this.futamIdo = futamIdo;
