+++ main/java/KoltsegDecorator/IKoltseg.java
@@ -3,5 +3,5 @@ package KoltsegDecorator;
 public interface IKoltseg
 {
     String getKoltsegNem();
-    double getKoltseg(int nevErtek);
+    double getKoltseg(long nevErtek);
 }
