+++ main/java/KoltsegDecorator/Koltseg_ErtekPapirSzamla.java
@@ -12,7 +12,7 @@ public class Koltseg_ErtekPapirSzamla extends KoltsegDecorator
     }
 
     @Override
-    public double getKoltseg(int nevErtek)
+    public double getKoltseg(long nevErtek)
     {
        return alapKoltseg.getKoltseg(nevErtek) + nevErtek*ertekpapirSzamlaDij;
     }
