+++ main/java/KoltsegDecorator/Koltseg_Riaszto.java
@@ -14,7 +14,7 @@ public class Koltseg_Riaszto extends KoltsegDecorator
     }
 
     @Override
-    public double getKoltseg(int nevErtek) {
+    public double getKoltseg(long nevErtek) {
         return alapKoltseg.getKoltseg(nevErtek) + riasztoBeszerzesiKoltseg + riasztoBeszerelesiKoltseg + riasztoUzemeltetesiKoltsegEvente;
     }
 }
